feat: move files to folders
